requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic-settings==2.2.1
google-cloud-pubsub==2.21.1
google-cloud-storage==2.16.0
pyarrow==16.0.0
redis==5.0.3
pytest==8.1.1